my_financial_plan_backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
mysql-connector-python
pydantic
python-multipart
passlib[bcrypt]
python-jose
alembic
python-dotenv